+++ packages/wix-ui-framework/src/cli-commands/export-testkits/typings.d.ts
@@ -1,0 +1,13 @@
+import { Process } from '../../typings.d';
+
+export interface Options {
+  definitions?: string;
+  components?: string;
+  exportSuffix?: string;
+  exportCaseStyle?: string;
+  template?: string;
+  output?: string;
+  factoryName?: string;
+  uniFactoryName?: string;
+  _process: Process;
+}