+++ packages/wix-ui-framework/test/cli.spec.ts
@@ -1,5 +1,6 @@
 import * as path from 'path';
 import * as nixt from 'nixt';
+import * as cista from 'cista';
 
 import * as packageJson from '../package.json';
 
@@ -52,15 +53,26 @@ describe('wuf', () => {
       });
     });
 
-    it('should ask for component name', done => {
-      nixt()
-        .expect(({ stdout }) => {
-          expect(stdout).toMatch(/Component name \(PascalCase\).*Test/);
-        })
-        .run(cli('generate'))
-        .on(/Component name/)
-        .respond('Test\n')
-        .end(done);
+    describe('--component-name', () => {
+      it('should prompt for component name by default', done => {
+        nixt()
+          .expect(({ stdout }) => {
+            expect(stdout).toMatch(/Component name \(PascalCase\).*Test/);
+          })
+          .run(cli('generate'))
+          .on(/Component name/)
+          .respond('Test\n')
+          .end(done);
+      });
+
+      it('should not prompt for component name given --component-name flag', done => {
+        nixt()
+          .expect(({ stdout }) => {
+            expect(stdout).toMatch(/Generating.*<Testable\/>.*component/);
+          })
+          .run(cli('generate --component-name Testable --force'))
+          .end(done);
+      });
     });
   });
 
@@ -94,7 +106,7 @@ describe('wuf', () => {
         nixt()
           .expect(({ stderr }) => {
             expect(stderr).toMatch(
-              /Error: Definitions file does not exist at "whatever.js"/,
+              /Error: Definitions file does not exist at ".*whatever.js"/,
             );
           })
           .run(cli('export-testkits --output test --definitions whatever.js'))