+++ packages/wix-ui-framework/CHANGELOG.md
@@ -11,6 +11,21 @@ Types of changes:
 1. **Fixed** for any bug fixes.
 1. **Security** in case of vulnerabilities.
 
+# 2.4.1 - 2019-06-18
+## Fixed
+- `wuf generate` - fix `codemods` flag to correctly resolve relative path
+
+# 2.4.0 - 2019-06-17
+## Added
+- `wuf export-testkits` - new `components`, `exportSuffix` & `exportCaseStyle` flags
+
+# 2.3.0 - 2019-06-12
+## Added
+- `wuf generate` - new `--output` flag to set generated files path
+
+## Changed
+- `wuf generate` - change `--codemods` flag to be optional. By default no codemods will be run
+
 # 2.2.0 - 2019-05-28
 ## Added
 - `wuf export-testkits` - new `factoryName` & `uniFactoryName` flags